# backend/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/stock/out": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock Out",
                "operationId": "3fa7336cd8156cf39be4106d24a33374",
                "requestBody": {
                    "description": "All Request Body For Stock Out",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "qty": {
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "date_out": {
                                        "type": "string",
                                        "format": "date",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stock/out/list": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock out list",
                "operationId": "0db46c92355dc993cb4c5cbc5e08422e",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stock/out/update/{id}": {
            "put": {
                "tags": [
                    "Stock"
                ],
                "summary": "Update Stock Out",
                "operationId": "360e03f454cdc0f0789f68f819582c66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id For Stock Out",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "All Request Body For Stock Out",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "qty": {
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "date_out": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-26"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/suplier/create": {
            "post": {
                "tags": [
                    "Suplier"
                ],
                "summary": "Create Suplier",
                "operationId": "f84c330890a0aea2b0d7d7aa326dca36",
                "requestBody": {
                    "description": "All Request Body For Suplier",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PT komputer Jaya"
                                    },
                                    "contact_person": {
                                        "type": "string",
                                        "example": "Budi Tabuti"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "082233712925"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "budi@gmail.com"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl Flamboyan no 4"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/suplier/delete/{code}": {
            "delete": {
                "tags": [
                    "Suplier"
                ],
                "summary": "Delete Suplier",
                "operationId": "2cb4e173c2f7fba46a6f3ddb99d3f48f",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Code For Suplier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}